Refactor messages screen in app
