+++ apps/app/lib/src/features/messages/presentation/message_scroll_view.dart
@@ -1,0 +1,72 @@
+import 'package:app/src/common_widgets/list_items_builder.dart';
+import 'package:app/src/features/messages/application/messages_service.dart';
+import 'package:app/src/features/messages/domain/message.dart';
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:intl/intl.dart';
+import 'package:jiffy/jiffy.dart';
+
+class MessageScrollView extends HookConsumerWidget {
+  const MessageScrollView({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return SizedBox(
+      height: 300.0,
+      child: Consumer(builder: (context, ref, child) {
+        final messages = ref.watch(messagesServiceProvider);
+
+        return ListItemsBuilder<Message>(
+          data: messages,
+          itemBuilder: (context, model) => MessageListTile(
+            title: model.title,
+            message: model.message,
+            dateTime: model.dateTime,
+          ),
+        );
+      }),
+    );
+  }
+}
+
+class MessageListTile extends StatelessWidget {
+  const MessageListTile(
+      {super.key,
+      required this.title,
+      required this.message,
+      required this.dateTime});
+
+  final String title;
+  final String message;
+  final IconData icon = Icons.message;
+  final Color color = Colors.red;
+  final DateTime dateTime;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+        margin: const EdgeInsets.only(top: 15, left: 10, right: 10, bottom: 0),
+        decoration: BoxDecoration(
+          color: Colors.white,
+          border: Border.all(color: (Colors.red[800])!, width: 2.0),
+          borderRadius: BorderRadius.circular(14.0),
+        ),
+        child: Card(
+            margin: EdgeInsets.zero,
+            color: Colors.white,
+            child: ListTile(
+                leading: Icon(icon, color: color),
+                title: Text(
+                  title,
+                  style: TextStyle(color: color, fontWeight: FontWeight.bold),
+                ),
+                subtitle: Text(message),
+                trailing: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    crossAxisAlignment: CrossAxisAlignment.end,
+                    children: [
+                      Text(Jiffy.parse(dateTime.toString()).fromNow()),
+                      Text(DateFormat('hh:mm a').format(dateTime))
+                    ]))));
+  }
+}
